refactor(cache): add cache module and add price to product object
- Add cache module.
- Add price to product objects.
- Progress bar.
- Styles.

+++ data/src/main/java/com/example/data/entities/Product.kt
@@ -12,5 +12,5 @@ data class ProductEntity(
     @SerializedName("thumbnail")
     val thumbnail: String? = null,
     @SerializedName("price")
-    val price: Int? = null
+    val price: Double? = null
 )
